scripts/requirements-x-dms.txt
# Python deps for scripts/fetch_x_dms.py
requests>=2.28.0
requests-oauthlib>=1.3.0
python-dotenv>=1.0.0
